+++ apl-lms-waybill-outstorage-pojo-master/apl-lms-waybill-outstorage-pojo/src/main/java/com/apl/lms/waybill/outstorage/pojo/vo/WaybillWaitOutstorageInfoVo.java
@@ -80,28 +80,28 @@ public class WaybillWaitOutstorageInfoVo implements Serializable {
     private String outChannelName;
 
     @ApiModelProperty(name = "inActualWeight" , value = "收货实重")
-    private BigDecimal inActualWeight;
+    private Double inActualWeight;
 
     @ApiModelProperty(name = "inVolumeWeight" , value = "收货体积重")
-    private BigDecimal inVolumeWeight;
+    private Double inVolumeWeight;
 
     @ApiModelProperty(name = "inChargeWeight" , value = "收货计费重")
-    private BigDecimal inChargeWeight;
+    private Double inChargeWeight;
 
     @ApiModelProperty(name = "cargoType" , value = "入包裹类型")
     private String cargoType;
 
     @ApiModelProperty(name = "outActualWeight" , value = "出货实重")
-    private BigDecimal outActualWeight;
+    private Double outActualWeight;
 
     @ApiModelProperty(name = "outVolume" , value = "出货体积")
-    private BigDecimal outVolume;
+    private Double outVolume;
 
     @ApiModelProperty(name = "outVolumeWeight" , value = "出货体积重")
-    private BigDecimal outVolumeWeight;
+    private Double outVolumeWeight;
 
     @ApiModelProperty(name = "outChargeWeight" , value = "出货计费重")
-    private BigDecimal outChargeWeight;
+    private Double outChargeWeight;
 
     @ApiModelProperty(name = "outStatus" , value = "出货状态 0未加入 1已加入出货列表 2已出货")
     private Integer outStatus;
